+++ Assets/Scripts/ParalyzeTroop.cs
@@ -80,7 +80,7 @@ public class ParalyzeTroop : Troop {
         }
     }
 
-    public override void Damage(int dmg)
+    public override void Damage(float dmg)
     {
         if (!hasShield)
         {
